fix: parsing error on the switcher (#1809)
Currently the `Switcher` uses a `div` inside of an `span`. We do control the inner `div` but the `span` comes from the `Dropdown` component of **Office Fabric** so the direct way to solve this is to change the `div` into a `span`. Also, by doing this, there is no need to change css.

I updated the tests and added some missing ones.

* reduce duplication
* Change div for span to fix parsing errors