feat(electron-update): Use script to update electron package's checksum after signing (#1824)
Our electron update mechanism relies on a latest.yml metadata file to provide the package file name, version, and an accompanying checksum to validate the downloaded file. Currently, this metadata file is generated alongside the corresponding package file. In the build loop, we then sign that package file. This modifies the package, thus rendering the previously generated checksum invalid. This PR introduces a script to recalculate the checksum and update the latest.yml file accordingly. It then puts the script immediately after the signing steps for mac and windows.