feat(electron-update): Add task to prepare & publish electron artifact (#1828)
Currently, we publish an artifact for the electron build that includes both the packed and unpacked versions of the product for each platform. This is useful for testing purposes, but we only need the packed version for the actual release process. This PR adds a build step that copies only the files we need to release into a separate directory for use in the release pipeline. The artifact is then published as usual.