chore: make CSS module class names consistent between dev and prod (#1811)
This is a continuation of @waabid 's #1570. I've addressed the feedback that was there and fixed the e2e tests.

The most important bit of change is in `webpack.config.js`. Most of the other changes are either:
* Fixing pre-existing cases where we re-used class names across different modules. In particular, automated-checks-view and device-connect-view and the assorted cards report components did a lot of this between each other.
* Replacing e2e tests' use of classnames with `data-automation-id` attributes (a few web e2es and almost all electron e2es)

There is also a change in `app-controller.ts` that forces the electron E2Es to test against the minimum supported window size; this fixed an issue that came up during debugging where some test failures were happening in a way that reproed on build agents but not on dev machines.

Out of scope: updating all web CSS selectors used by E2Es to use `data-automation-id` instead (this would be a good idea to do to pre-empt tests breaking when moving more components to CSS modules, but leaving it for a separate PR)

* chore: css modules will always use the hash value for generating class names

* Consolidate common webpack bits
* Update e2e snapshots to avoid CSS module hashes
* Fix half-finished comment on normalizeCssModuleClassNames
* Undo irrelevant overview-heading unit test change
* Fix ordering of import vs copyright
* Fix electron e2es by using data automation IDs throughout
* Fix lint issue
* Update unit tests to account for data automation id attributes
* Fix device connect port entry test to use new automation id
* Add a wait for selector to attempt to address the build agent specific e2e issue
* Add quotes around data automation id values in selectors
* Fix test environment difference in electron e2e behavior
* Add intermediate state checks to flaky test case
* Force elecron e2es to run at the minimum window size consistently
* Fix non-:global cross-module references in electron views
* Update snapshots