fix: move interactive no-drag styles to electron root (#1878)
A recent change in an scss file caused the interactive elements on the telemetry dialog in the electron product to stop accepting mouse interaction. This happens because -webkit-app-region: no-drag was not set on those elements. Even though the telemetry dialog is a child of the device connect view, it actually renders as a sibling. This PR moves the -webkit-app-region: no-drag properties for buttons and inputs out to apply globally so that we don't have to worry about similar problems in the future.