refactor: promisify switch to tab (#1919)
This is part of a long-run refactor to promisify all (possible) `chrome.*` calls.

In this case, promisifying `switchToTab`.

`switchToTab` actually have a little logic inside (so not strictly an adapter function): 1) set the tab to active; 2) set the window to focused. Simple enough but doesn't really match the Adapter Pattern. As a future work (so out of scope for this PR) we may want to extract that bit of logic to, says, `TabSwitcher` class (or plainly a function).

* add update tab & window public api
* add promise-based switch to tab api
* update inspect-action-creator to use promise-based api
* update tab-action-creator to use promise-based api
* update details-view-controller to use promise-based api
* update target-tab-controller to use promise-based api
* remove callback-base api
* rename function
* fix tslint errors
* true up name, remove empty comment line
* fix typo

Co-Authored-By: lisli1 <48259897+lisli1@users.noreply.github.com>