fix: scanning spinner is not announced on assessment's automated checks (#1926)
Currently, Assessment -> Automated checks shows a "scanning" spinner while we run axe. We are using Office Fabric `<Spinner>` which does not get properly announce (it does not get announced at all).

In this PR, we are switching to use `<ScanningSpinner>` to get a proper announcement. This is the same component we are already using in the new cards UI (so it's already being use on electron app and fast pass -> automated checks)