refactor: use css grid to layout diagnostic view toggle (#1968)
Currently, the composition between the ad-hoc tools panel and the 5 different diagnostic view toggles is based on office ui fabric react grid. This make for a lot of boiler plate `<div>` elements, which contains `ms-Grid`-related classes. This are the classes you need to use if you want to use OF grid but that's not really necessary as we can use css grid and move more of the layout details to scss.

As part of this PR:
- convert `DiagnosticViewToggle` to use css-module
- convert `DiagnosticViewToggle` to use css grid
- update snapshots

I tried to make the styles to match the original layout as much as I could but there may be small differences.

* using css grid instead of office fabric grid
* clean up old, unused css
* fix test description string