feat(office-fabric): udpate tooltip host content (#1975)
After updating to office fabric 7, the Assessment in progress dialog got a layout-changing behavior when the user hover over the second target page link (this user story has more details 1663263).
This happen because we're mixing inline and block elements inside the dialog content. In order to fix this, I forced the <Link> component to render as an anchor tag (which is an inline type of element).