feat(office-fabric): fix switcher at announcement (#2002)
Several pieces of information to explain this change, bear with me:

The switcher AT announcement includes:
- tab-navigating to the dropdown.
- use arrow keys to change selection (with the dropdown to expand)
- expand the dropdown and use arrow keys to navigate the options.

Currently on **AI-Web prod 2.13.1** ATs announce the switcher **good enough** (there is just so much we can do to provide good aria values, after that, it's up to each AT to use that information and there are small and big differences between them.)

Currently on master, we have lost the proper announcement of the selection changes (by using arrow keys without expanding the dropdown). This is due to what we render for each option. We are using a wrapper `<span>` element so we can have icon + text. This wrapper has `aria-hidden="true"` which makes the ATs not to read it.

The fix then is as simple as remove this attribute from said element.

I have tested this change considering:
- browsers: (new) edge & chrome (they do behave different oddly enough)
- AT: Narrator (the best output), JAWS & NVDA

* udpate alert text
* remove aria-hidden to fix AT announcements