fix: tab url change spamming telemetry event (#2012)
Currently on prod 2.13.1 (and master) the extension is sending a telemetry event every time the user navigates any tab to a different url. The original purpose of the telemetry event was to gather data to understand how often do our users reuse a tab they where scanning but the implementation did not account for this last part, i.e. we're not checking if the tab url change was trigger on a tab the user was actually scanning.

This missing check makes the extension to spam this event and we have seen an huge spike on telemetry events which may make us hit a limit on the amount of events/data we can send to app insights per unit of time.

Given the 2 paragraphs bellow, we are removing this event temporarily until we: 1) decide we actually need it; and 2) design a way to do this properly (i.e. no spamming, event sent only under the right circumstances)