chore: log version when publishing ai-web (#2031)
While investigating longer-than-usual update times of our canary extension on the chrome web store, I noticed that our release do not log the version we're using when uploading to the store. 

This makes it harder to track which release run was the one that actually upload an updated version of the extension last, basically because our release does not fail if the store respond with an error, as in the case of uploading a new version while the extension is under "pending review"; in that case, we get this message:

```bash
Error on uploading (open) with message "Application error (6): hbcplehnakffdldhldncjlnbpfgogbem pending review so can't be edited". Raw response: {"kind":"chromewebstore#item","uploadState":"FAILURE","itemError":[{"error_code":"CLIENT_ERROR","error_detail":"Application error (6): hbcplehnakffdldhldncjlnbpfgogbem pending review so can't be edited"}]}
```

On this PR, just adding a log so we can actually see the version being used.

**Future work**: 
- we may want to change our logic to either fail when we get this message (that'll make it clear which release are actually successfully uploading an update)
- we may want to check extension status before uploading
- we may want to make sure the release has some sort of cooldown period (historically, we know that canary usually takes 15~20 minutes to update; it doesn't make sense to run a release if the last one was only 5 minutes ago)

* logging version value of ai-web extension when publishing
* remove unused require