fix: rule group header loses focus indicator while using AT (#2042)
There is something about how Office Fabric handle focus visibility state when the user interacts with the webpage. There seems to be some rules around how to change that state base on the type of element and the type of interaction (basically, keyboard vs mouse).

In our case, when the user uses the keyboard to collapse/expand the rule header group, office fabric simply set the focus visibility state to hidden.

On the other hand, Office Fabric expose a `setFocusVisibility` utility function we can call to force the state. 

**Note** The report is not affected by this bug (only AI-Web and AI-Android). The main reason is the report is mostly static html, so there is no office fabric javascript running on the report.

* add type
* set focus state to visible when the user uses the keyboard to expand/collapse a rule group