refactor: root-container to use withStoreSubscription (#2088)
Currently, AI-Web follows a pattern of wrapped the highest level component using `withStoreSubscription`. This way, all the store listening and reacting to updates is encapsulated on one component, leaving the high-level component free to take care of other responsibilities.

On AI-Android, this is not the case, as `<RootContainer>` explicitly handles listening to the stores (through an stores hub instance, but still).

On this PR: refactoring `<RootContainer>` to  use `withStoreSubscription` and get rid of all the store-listening-and-updating logic. This change will make all **"containers"** to be more alike and also -I think- it will help to enabling high contrast theme on AI-Android.

**NOTES** no UI/UX changes from this PR

* refactor: root-container to use withStoreSubscription
* fix lint errors