refactor: is-mac-os class on AI-Android (#2092)
Currently, AI-Android uses different css styles for the window title bar base on which OS the app is running on (mac os vs others). The different styles are being applied at the `<WindowTitle>` level, not at a higher level (`<body>` element for instance) like in the case of `'high-contrast-theme'`.

Working on adding the **settings panel** to AI-Android, I noticed there will be a need to do differential behavior for the settings panel component based on OS as well, so it makes sense to move and rename `.mac-window-title` from it's current placement to the `<body>` element, so any component that needs different styles based on OS can use the same class. 

Doing this matches what we do with high contrast theme, which is something we need to bring to AI-Android as part of the settings panel feature anyway. This PR will also help with that part.

**Notes** there is no actual UI/UX behavior changes, only how we apply class names and styles

* move window title files to their own folder
* add BodyClassModifier to add is-mac-os when is proper
* use BodyClassModifier on root container renderer
* udpate WindowTitle to not add is mac os class name to header element