refactor: popup header menu buttons (#2063)
While working on #2053, I tried to convert the related components to use CSS Modules but that turned out to be harder than it should. I think the main reason this happened is because of the way we decided to implement the buttons on the header (**gear** and **hamburger**):

1. **gear**: it's actually using `<Link>` and `<ContextualMenu>`
1. **hamburger**: it's using `<IconButton>` and `<ContextualMenu>`

This decision have some side effects:
1. Using two different component types makes it harder to style the components so they look the same.
1. Accessibility is not correct, as both elements are being announced as _plain_ buttons. There is no indication that activating either of the buttons will show the contextual menu.
1. Manually using `<ContextualMenu>` force us to have state-full components, which in general, tend to be more complex than stateless components.

This side effects seem to go away if we choose to use `<IconButton>` **and** to pass `menuProps` property when rendering the buttons. Passing `menuProps` effectively convert the elements into **menu** buttons, which means an AT, NVDA for instance, will announce the element something like this:
```
manage settings
menu button collapsed submenu
```
Note that **manage settings** is our own, custom 'aria-label' value, while the role **menu** and the state **collapsed** comes directly from using `<IconButton>` + `menuProps` property, there is nothing extra we need to do to achieve this.
Also, the new components will be stateless as the `<IconButton>` will take that responsibility now.

**NOTE**
I know this looks like a big PR. I did try to break it on smaller chunks but it was too hard to achieve at both the typescript and css level without doing improper things. I guess it could be easier to review commit by commit instead.

* full refactor of popup header menu buttons
* refactor gear-options-button-component. minor improvemnt on the test
* refactor/fix gear button in the details view
* fix button color and callout border on high contrast theme
* rename component
* rename component file
* rename component scss file
* rename component test file
* update launchpad e2e test snapshot
* fix selectors for the hamburger menu e2e test
* update selectors for gear menu button on the details view for e2e test
* update another, unrelated e2e snapshot
* udpate settings drop down e2e test
* rename props type to reflect renamed class
* rename nearItems to items
* rename function to match property