fix: unified/mac: support signed build checksums for latest.yml with multiple files (#2131)
#### Description of changes

Before #2126, all of the `latest.yml` files we produced contained only a single file, so it was okay for our checksum-updating script to assume it only had to update `files[0]`. Now that we have a case where multiple files can be present, this PR makes that update script a little smarter about updating checksums on a per-file basis. 

Verified locally against a single- and multi- file latest.yml

#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a" in the checkbox -->
- [n/a] Addresses an existing issue: #0000
- [x] Ran `yarn fastpass`
- [n/a] Added/updated relevant unit test(s) (and ran `yarn test`)
- [n/a] Verified code coverage for the changes made. Check coverage report at: `<rootDir>/test-results/unit/coverage`
- [x] PR title *AND* final merge commit title both start with a semantic tag (`fix:`, `chore:`, `feat(feature-name):`, `refactor:`). Check workflow guide at: `<rootDir>/docs/workflow.md`
- [n/a] (UI changes only) Added screenshots/GIFs to description above
- [n/a] (UI changes only) Verified usability with NVDA/JAWS