feat(android-settings): bring settings panel to AI-Android (#2120)
This PR hook up all necessary pieces to bring the settings panel (from AI-Web) into AI-Android.

There was only one `NullObject` needed to pull this off.

Styles needed to be adjusted for the `<Panel>` to not be render on top of the window title bar. The styles also account for the height difference when running AI-Android on MacOS.

After this PR is completed, bringing **high contrast mode** should be straight forward, although, we'll need to check if all related styles are being applied correctly and that nothing is left behind (i.e. elements that are not affected by high contrast, giving the app and mixed-style look)

**Notes**
In order to keep this PR small, e2e tests are consider out of scope of the PR, but not from the ~not-yet-active~ feature.

* add dummy gear icon settings button
* hook up settings button click handler
* hook up details view action creator
* hook up details view store
* hook up settings panel
* fix wrong type after rebase from master
* override styles to properly position the settings panel
* alphabetize imports
* update unit test for automated checks view
* fix settings panel unit tests
* Rename class
* Rename interface
* Rename class file
* Rename interface file
* rename test
* Rename exported const
* Rename file
* alphabetize imports