feat(android-settings): fix outcome chip color to respect high contrast theme (#2177)
Currently on master, out outcome chip is using `color: $neutral-90`, which is defined as `$neutral-90: var(--neutral-90)`. 
The issues with this is `--neutral-90` is not actually defined. 
The normal behavior of CSS custom properties (more info https://developer.mozilla.org/en-US/docs/Web/CSS/Using_CSS_custom_properties) make the outcome chip to set the color by using one of  it's ancestors color value, which mask the fact that `--neutral-90` is not defined. 
Another issues with this is: for some reason, on AI-Android the outcome chip does not react to high contrast mode being turned on (probably because there is no ancestor that set the color value following high contrast mode).

On this PR:
- I used `$primary-text` value to set the outcome chip color. This matches what we currently see on the UI.
- Remove all references to **neutral-90** (var and custom property) so we don't end up using it somewhere else by mystake