feat(android-settings): fix rule description hover state text color on high contrast (#2192)
Currently, hovering over a rule group on AI-Android makes the description to change text color to something closer to black, which makes it unreadable. This is due to a color: unset; value we are setting specifically for the hover state plus some missing color value on the parent/ancestor of this element (color is an inheritable property, color: unset; makes css to look up to the parent of an element and up the hierarchy until it finds a color or uses a default value).

In this PR: using an explicit color for the hover state which fixes AI-Android and leaves AI-Web and the report with the same, original behavior.

Current master: ImageViewName rule hovered over, where the description is not visible due to the text using black.