feat(android-settings): add body color value (#2196)
Currently, we don't have a body color value on AI-Android. This makes the color value to be black (the default value). This default value does not change when we enable high contrast mode which make the text practically invisible against the black background.

This is easily fixed by adding a explicit color value for the body. Using `$primary-text` for this.