fix: how do I connect my device? link role (#2219)
Office Fabric `<Link>` component uses a native `<button>` element when there is no `href` prop pass to it.
We are using the `<Link>` component but we can't pass the `href` because we want AI-Android to use the native, OS default browser to open the link. This works by us passing a `onClick` function that wraps the `shell` object from `electron` to call into the OS to open the link.

Due to the previously exposed, it's kind of hard to use Office Fabric component **and** get a `<a>` element, so instead. Thus, the approach taken on this PR is to force the `<Link>` to use `role='link'`.