feat(android-settings): cloneDeep on UserConfigurationStore (#2229)
This is a follow up of #2199. 

The previous approach allows us to change the behavior of `getState` when building instances of any store. The **any** store part is the critical part here. We do want to make our states immutable and we have #2228 to track that work but following the original approach cause a series of issues on AI-Web so instead of opening the door and facilitate this change on any and all the stores, we'd like to take a safer approach and update only the offending store (on AI-Android, that would be `UserConfigurationStore` where the `enableHighContrast` flag is store).

* override getState to always use cloneDeep
* remove stray .only