feat(android-settings): update command bar to fix a11y issues (#2241)
Currently, we are using `<CommandBar>` from Office Fabric on the Automated Checks View to show the start over button and the settings gear. The component has some accessibility issues capture on [OF#11832](https://github.com/OfficeDev/office-ui-fabric-react/issues/11832). 

I fiddle around a little bit with the `<CommandBar>` and the `<FocusZone>` components to see if we could get a walk-around and still use the Office Fabric component but I didn't find anything promising and already hit my time box.

So instead, we are falling back to the approach we have on AI-Web with the export report and start over buttons: basically, the buttons live inside a simple `<div>` and the user can tab between them normally. There is no special announcement from the AT (like amount of items in the "command bar") but everything is keyboard-navigable and discover-able.

I also used the opportunity to update the buttons style to match the figma mock with the general style for command buttons.

* stop using command bar; update command bar to use flex and simple divs
* remove unneeded spacer div
* update snapshot