feat(android-settings): update how do I connect link colors (#2251)
Currently, to render the **How do I connect to my device?** link, we directly use `<ElectronLink>` component, while for the links on the settings panel, we pass a dependency called `LinkComponent`. Now, `<ElectronLink>` can (and is) be passed as the instance for the dep `LinkComponent` but we should have only one way to do this so, defaulting to use the dep instead of have a hard dependency on `<ElectronLink>` (this helps with open/close principle).

Additionally, we have override the css styles for the link and now the **How do I connect to my device?** link looks different from the ones on the **Settings Panel**. 

In this PR, we make both links look the same (using the **Settings Panel** as the style we want to use). I'm following the **Basic buttons** style guideline from figma, specifically, the hover state, to set the new colors for the links.

Note that instead of doing this on the css, I'm updating `HighContrastThemePalette`, which means this color will apply everywhere as long as we don't have css overrides. This is the case for AI-Android, so now all our links will use this colors. I checked on AI-Web and the new yellow color use on the link hover state is not use, which means we have overrides. I think we should favor using the palette instead of css overrides as this is the intended way (although not 100% completed, nor perfect) to do this on Office UI Fabric React components.

* using LinkComponent from the dependencies instead of hardcoding ElectronLink
* update HC palette, remove css overrides