feat(android-settings): make unified command bar a region landmark (#2257)
#### Description of changes

This PR updates the newly redone unified command bar to be a "region" landmark named "command bar", per the figma (`<section>` elements are the semantic HTML equivalent for the `region` role). We think eventually a "menubar" role would be more appropriate, using this as a stopgap while we wait for https://github.com/OfficeDev/office-ui-fabric-react/issues/11832

No visible UI changes. Verified reading behavior in JAWS (it makes the region navigable with "r").

#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a" in the checkbox -->
- [n/a] Addresses an existing issue: #0000
- [x] Ran `yarn fastpass`
- [x] Added/updated relevant unit test(s) (and ran `yarn test`)
- [x] Verified code coverage for the changes made. Check coverage report at: `<rootDir>/test-results/unit/coverage`
- [x] PR title *AND* final merge commit title both start with a semantic tag (`fix:`, `chore:`, `feat(feature-name):`, `refactor:`). Check workflow guide at: `<rootDir>/docs/workflow.md`
- [n/a] (UI changes only) Added screenshots/GIFs to description above
- [x] (UI changes only) Verified usability with NVDA/JAWS