feat(android-settings): add e2e tests for the settings panel (#2297)
Adding e2e test for the settings panel on AI-Android, including accessibility issues scan in default and high contrast theme.

The general strategy was to bring the same scenarios on the settings panel e2e test from AI-Web but translated to AI-Android.

**Note**
We have the idea to update the app controller to use the user config controller instead of hitting buttons in the UI for dismissing the first time telemetry dialog. This will lead to some changes to other e2e as well as adding a couple more e2e test. I left that for follow up PR to keep this one smaller.

* add e2e tests for the settings panel
* add data-automation-id attribute to gear button
* add data-automation-id attribute to gear button
* update settings button selector
* skip test for system/default HC mode check