refactor: action hub dependency on action-creator test (#2314)
We need an instance of `ActionHub` to run the tests for the `ActionCreator`. We are creating that instance manually, assigning each property with a mock or a null value. This force use to update this test every time we add a new "actions" property to the `ActionHub` even if that new property does not affect the `ActionCreator` test.

By using casting (as in `as ActionHub`) we can avoid this, remove the null properties and only assign mocks to what we actually need on the test.