refactor(side-panel): add side panel actions to open settings panel (#2313)
Currently, **AI-Web** has 4 "side panels" (**AI-Android** is only using the Settings panel at this point): 
- Settings
- Preview Features
- Scoping
- Add a Failure Instance
- Info & Examples. 

All of the side panels except Add a Failure Instance are shown/hidden by updating the state on the `DetailsViewStore` by invoking specific actions:
- openPreviewFeatures
- openScopingPanel
- openContentPanel
- openSettingsPanel

There are also 4 actions to close each of those panels.

This could be improved by handling the side panel state in a more generic way, for example, by having a unique open side panel action which takes the type of panel we want to show. The current show panel logic ensures that only one panel is shown at most at any time, which means, using one action to opening is 100% feasible without changing the store logic to much.

This PR is the first one in the path of applying this idea. It only change how we open the settings panel, setting the precedent of how to update the other open panel logic. It also does most of the piping we need in order to do this refactor making the next refactors potentially smaller.

**Note**
Along the same line, we may want to change the state values we keep on the store, from this:
```javascript
currentPanel: {
    isPreviewFeaturesOpen: false,
    isScopingOpen: false,
    isContentOpen: false,
    isSettingsOpen: false,
}
```
to something less verbose (like just storing the type of the panel we are showing or `'None'` if we are not showing any panel).

* add side panel type
* add side panel actions
* update the store to handle openSidePanel action being invoked
* improve side panel to store key dictionary type
* remove stray empty commented  line
* add sidePanelActions to ActionHub
* make side panel action a mandatory parameter
* using openSidePanel to open the settings panel (instead of openSettingsPanel)
* remove old openSettingsPanel action