feat(threat-model): enable tslint rule for no dangerous html property (#2358)
As part of our threat model follow up, we need to handle our usages of `dangerouslySetInnerHtml`. That means remove them or somehow account for them (in the sense of: let's not add more usages and make it clear where and why we are using it).

There is a microsoft contrib tslint rule to catch uses of `dangerouslySetInnerHtml`. In this PR I'm enabling that and grandfathering the current usages of this property through our code.

I do thing there is more we can do to actually remove this usages, so we can either have this PR as a backup plan, or we can go ahead (so we have the tslint rule and no more usages are introduce) and fix what we can later (still during this feature).