feat(threat-model): introduce target page inspector (#2382)
As part of our threat model follow up we need to assess our usage of javascript `eval`(which we don't have any and we do have a tslint rule to catch this). While looking at our code base, I found we are using `chrome.devtools.inspectedWindow.eval` (this allow the user to inspect an element we marked as failing an automated check, directly from the target page's details dialog).

While we are using this `eval` API only to call `inspect`, we are implementing this in a way we can pass any script to `eval` (not only `inspect`) which is not necessary and also a opens the door for some security risks.

In this PR:
- Change the API so the user cannot pass just any script; the new API is specific for inspect, the user can pass a selector
- Run some basic sanitization on the input (the selector) as a malicious attack can still try to use this user input to try and run arbitrary code.

**Discussion**
This new API is not bullet proof; it's not close to be perfect neither. This is a starting point in case we start hearing about security concerns with concrete examples on how to exploit this on our extension. 

* add target page inspector
* update InspectHandler to use TargetPageInspector
* remove executeScriptInInspectedWindow
* alphabetize imports
* improve comment grammar
* improve how we test the happy path
* update test cases with injection attempts. remove unneeded test
* update our injection attempt selector test cases
* simpify test cases and add a couple of new ones
* add exception for fake, not explicitly used function