refactor: multiplexing telemetry client (#2397)
Currently, we use some sort of strategy pattern when it comes to our telemetry clients: if we build without an app insights instrumentation key, we use a `NullTelemetryClient`, which do not send any telemetry to external services; if we do have an instrumentation key (as in the case of a released build) we use an `AppInsightsTelemetryClient` to actually send telemetry.

In both cases, the telemetry client uses a `TelemetryLogger` so we can log any event to the background console (mostly for debug/validation purposes).

This works just fine! But, I'm planning on moving the "local console telemetry" to our debug tools page. In this case, our debug tools will be just another "sink" for telemetry events, one that we may always want to send events to, while leaving the app insights sink with the same current behavior (base on the instrumentation key being present or not).

I decided to follow @dbjorge comment and implement a simple container to pass the telemetry client calls to all of them (a simple array and simple iteration through it).

On this PR:
- Introduce the `MultiplexingTelemetryClient`, to wrap the original `TelemetryClient` implementation (so no changes on those classes), and still allow to create the chain
- Update the `TelemetryClientProvider` to use the new chain of responsibility
- Rename `NullTelemetryClient` to `ConsoleTelemetryClient` to better reflects its responsibility
- Remove the usage of the `TelemetryLogger` from `AppInsightsTelemetryClient`; the main reason to do this is: the chain always have the `ConsoleTelemetryClient` as it's first link of the chain.

* add ChainnedTelemetryClient
* update telemetry client provider to return a chaiined telemetry client
* remove the telemetry logger usage from app insights telemetry client
* rename NullTelemetryClient > ConsoleTelemetryClient
* rename null-telemetry-client > console-telemetry-client (filenames)
* remove unneeded import
* rename const
* add multiplexing telemetry client
* update telemetry client provider to use multiplexed telemetry client
* remove chainned telemetry client
* rename files to match class name
* asserting amount of clients