refactor(side-panel): add generic onOpenSidePanel (#2420)
Currently, our `DetailsViewActionCreator` contains an specific function to open the settings panel. In order to facilitate future refactors to change how we open/close the other settings panels (PreviewFeatures, Issue Filing, etc), it would be nice to have a more generic way to open a side panel, thus, minimizing the amount of changes we need to make to "onboard" the other side panels.