feat(android-report): Wire up feature flag controller (#2456)
In order to facilitate feature work, it would be nice to have feature flags to hide/show in-progress features. We want to be able to hide the "Export" report button using a feature flag but currently, AI-Android does not have this kind of control.

On this PR we are making the `FeatureFlagsController` available to the devs. At this point, there is no UI work for this, so the controller will be available through the window console (dev tools).

Additionally, we remove the usage of the `RiggedFeatureFlagChecker` (it was being used only on AI-Android, specifically on the `TelemetryLogger`) and update how we initialize the `TelemetryLogger` which now uses the real `FeatureFlagsController`.

Finally, we tested the feature flag data is properly persisted to indexed db every time we change a feature flag and also, it's load from indexed db during the app initialization (so feature flag state is the same across app restarts)

* Hook up feature flag controller and dependencies
* Using the real FeatureFlagController on TelemetryLogger
* remove RiggedFeatureFlagChecker

Co-authored-by: kepownel@microsoft.com <kepownel@microsoft.com>