tools: add debug tools telemetry client and listener (#2445)
This PR introduces most (maybe all) of the piping needed to send and consume telemetry events on the debug tools.

The implementation relies on chrome API's that allows us to more explicitly establish a communication channel between different context (in this case, the background and the debug tools page). This allow for more fine-grained control.

There is no UI being added on this PR so to keep it more or less simple. The way to check it is working is to open the debug tools tab and inspect the page; the console will show all the telemetry events being trigger by the extension.

* add debug tools telemetry connection name
* add DebugToolsTelemetryClient
* add TelemetryListener
* update DebugToolsTelemetryClient to also send ApplicationTelemetryData as part of the custom properties
* refactor TelemetryClientProvider to accept baseClients array and allow for different behavior on AI-Android and AI-Web
* clean up imports