refactor: use clearer naming for export format vs service (#2503)
#### Description of changes

This PR is strictly refactoring, no functionality changes.

Before this PR, we have types for the following:

* `ExportResultsType`: is it an assessment report or a fastpass report
* `ExportFormat`: is the user downloading the report as a file, or sending the report to codepen

I found this confusing when rereading the code, since the former concept feels much more like a "format" to me than the latter concept. The types' naming were also a little inconsistent with other types in this space, which most commonly used use `ReportExport` as the prefix rather than `ExportResults` or `Export` (eg, the `report-export` top level folder name for the `ReportExportService` that `CodePenReportExportService` implements).

This PR renames those two types to `ReportExportFormat` and `ReportExportServiceKey`, respectively. It updates the related variable names, **with the intentional exception of the property name sent to telemetry**, to avoid breaking correlatability with past telemetry.

There are also a few more tightly scoped renames in the `code-pen-report-export-service`:
* Renamed files from `codepen` to `code-pen` prefix to match type `CodePen`-like type names ("CodePen" is the canonical way to capitalize the service)
* Updated a `const` from PascalCase to camelCase for consistency
* Used `type` instead of `interface`
* Standardized a few other report export service related types on `ReportExport*` instead of `Export*` type names

#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a" in the checkbox -->
- [n/a] Addresses an existing issue: #0000
- [x] Ran `yarn fastpass`
- [x] Added/updated relevant unit test(s) (and ran `yarn test`)
- [x] Verified code coverage for the changes made. Check coverage report at: `<rootDir>/test-results/unit/coverage`
- [x] PR title *AND* final merge commit title both start with a semantic tag (`fix:`, `chore:`, `feat(feature-name):`, `refactor:`). See `CONTRIBUTING.md`.
- [n/a] (UI changes only) Added screenshots/GIFs to description above
- [n/a] (UI changes only) Verified usability with NVDA/JAWS